Adds confidence metric to goto expressions.
This commit adds a `Confidence` enum, which represents the degree of
confidence the generator of the goto expression has that the expr
accurately represents the semantics of the underlying system being
represented in goto.  All expressions start with confidence `Default`,
which can be overriden using the `.with_confidence()` fluent builder.

This commit also adds a visitor which replaces low confidence Exprs
with an `assert(0); expr` statement expression. This ensures that
if low confidence expressions are dynamically used during a proof,
then CBMC will raise an assertion violation.

This visitor is currently hard-coded. Using a proper visitor pattern,
and possibly moving the assertion from the expression to the
containing statment is future work.