Change RMC to use rustc modified monomorphizer #295
Instead of keeping our own copy of the monomorphizer, we would like to
start using rustc monomorphizer as is. This is the first step towards
that goal.

For now, we will use the upstream version of the monomorphizer with a
few small changes. Use a query to skip monomorphization for rmc instead
of hooks.

This change, add a query to allow RMC to skip monomorphizing functions.
- Add a default query to codegen base that don't skip anything.
- Add a query implementation to RMC that checks our hooks.
- Skip monomorphizing the start function if --cfg=rmc

Issue #485 tracks the rest of the changes needed.