Remove is_rmc check from monomorphizer (#485)
We have changed rmc to always compile the target crate as a library.
This allow us to treat any extern / public function to be used as the
starting point of a proof. Furthermore, it allow us to remove the is_rmc
check that we were using to skip the rust start function generation.
The start function is only generated when we are compiling a bin crate.