Add script to format python files and format files (#584)
* Add script to format python files and format files

* Add messages to make it more user friendly

* Ignore 2 blank space issue for now in python format

We use the wrong convention to document functions. When enabling the
autopep8 it added 2 blank spaces between the function comment and the
function definition.

For now, we will disable the pep8 convention that sets the 2 blank space
before a function. This makes the code more readable. We should enable
this back once we fix the function comments.

* Apply suggestions from code review

Co-authored-by: Adrian Palacios <73246657+adpaco-aws@users.noreply.github.com>

* Fix trailing whitespace

Co-authored-by: Adrian Palacios <73246657+adpaco-aws@users.noreply.github.com>