Use rustc TyS as is (#42)
There is no need to access the kind and flags field of TyS directly.
We should be using kind() and flags() methods instead. Thus, I'm
removing the modification we made to the original rustc code to make
these variables public and fixing the code that relied on the direct
access.