Add any / any_raw to replace nondet.
Users will be able to control how RMC generate symbolic values to their
variables.

- `rmc::any::<T>()` will provide a safe way to generate values
that respect any invariant specified to type T (Given by the
implementation of Invariant type.)

- `rmc::any_raw::<T>()` will provide any bitvalue possible for the
memory layout of the type. This is an unsafe method and users must
ensure correctness after this call.

Note: I still need to add tests to the implementations of Invariant we
provide.