Revert change to rustc SwitchTargets type.
We have changed the visibility of one of the members of SwitchTargets to
have access to their values during codegen. I re-wrote the code to use
standard API so we don't need to keep using a forked version of the
compiler. Issue #42.

The only changes left are the ones to enable rmc codegen and a a few
options we enabled. Those should be handled once we have a new driver.