Add support to Cargo.toml metadata tables (#740)
Fixes #739.

Cargo rmc is able to extract flags from a toml configuration file. In
order to provide a good experience to users that want to add
configurations to their Cargo.toml, we are adding support to the
standard metadata tables supported by cargo.

We still support the [rmc] tables in cases users prefer to use a
different rmc.toml file, but this is discouraged for Cargo.toml files
since it will generate a warning every time a user runs cargo.