Fix kani-compiler logs + add json option
This change fixes #785. It allow users to configure the Kani logs via
KANI_LOG environment variable as well as via --log-level argument. Note
that the former allows finer grain control on a per crate configuration.

We also support the RUSTC_LOG variable to control the underlying rustc
packages. In order to control the rustc logging, we would need to match
the exact version of the tracing crate used by rustc. To avoid that,
we're keeping them separate.