Set kani-compiler verbosity correctly
As discussed previously, we would like to emit kani-compiler errors
and warnings by default. If user run kani with --verbose, this should
lower the kani-compiler logging level to info, and --debug will result
in lowering to compiler debug as well.

We are not changing the rustc logging level at this point. Users can
still configure it via RUSTC_LOG environmental variable.