Get rid of kani-rustc (#924)
This changes a few binaries in order to get rid of kani-rustc. Bookrunner, compiletest and cargo-kani invoke kani-compiler directly now. For our tools and regression scripts, I am keeping the assumption that this is a debug build. I removed that assumption for cargo-kani (but I haven't tested it yet for release build). :)

Unfortunately we still rely on the environmental variable hack to enable
cargo build. For now, we are moving it inside kani-compiler. :(