Fix config argument precedence and handling of CBMC options.
Cargo kani wouldn't work if the project Cargo.toml had --cbmc-args and
if the command line also included --cbmc-args.

The other issue is that the options in the project configuration had
precedence over the ones from the command line, which is misleading.
For example, if the Cargo.toml had a harness function defined to run by
default, the user would still run the default harness even if they ran:

```
cargo kani --function <different_harness>
```