Disable extra pointer checks (#1059)
* Add --enable-unstable option

Unstable options now require users to explicitly enable them. This is
important because the unstable options can be unsound or give spurious
CEX.

* Disable spurious pointer checks by default

We have been seeing a lot of spurious CEX related to pointer out of
bounds checks in arithmetic operations. It is actually valid rust in
some cases, and we can't differentiate the valid ones from the invalid
ones today. Thus, I'm disabling them by default.

Note that dereferencing invalid pointers is caught by different checks
and they should not be affected by this change.

This changes fixes #563 and #763.

Co-authored-by: Zyad Hassan <88045115+zhassan-aws@users.noreply.github.com>