Remove restriction to double fat pointer projection (#1065)
We were restricting projections to only go through a max of one fat
pointer dereference. This is not always the case, however, each
projection should only care about the outer most fat pointer.

Added tests to ensure this works as expected.

It looks like size_of_val is returning the wrong values for the
projection of traits. I'm adding fixme test cases for those here and
I'll create a new issue for them.