Inject runtime error for type mismatch in projection (#1086)
This is to mitigate potentially wrong code generation detected by our
type mismatch logic. Eventually we want to enable this assertion and
fail at runtime, but this currently affects the compilation of std.

Note: I had to fix the projection error handling for codegen_place.
This wasn't working if the failure was in the middle of the projection.