Truncated history.
Kani (originally 'RMC') used to be a fork of Rust.
This later changed, and this history truncation is the result of filtering out
Rust's history and some of its files from the Kani repo, to reduce repo size.
Some more information about this truncation can be found here:
https://github.com/model-checking/kani/issues/991

Kani originally pursued a rebase strategy for syncing with rustc and then
switched to merging, before finally using rustc as a library.
When this git history was truncated, this commit had rustc parent:
https://github.com/rust-lang/rust/tree/f8810ee171d1549cfb4082d4e1289704aad722e9
This commit reflects the last time the rebase strategy was applied.