Replace vector-map by linear-map dependency (#1120)
The `vector-map` crate indirectly depends on a vulnerable version of
rand_core. This crate hasn't been released for the past 2 years,
and there hasn't been any activity in their github either. I'm replacing
it by `linear-map` for now to fix the security warnings.

Note that linear-map last release was 6 years ago, so we may have to
remove it eventually. At least this crate only depends on serde.