Fix type mistmatch of types with projection
Structures with members that had type projections in their declaration
were being handled incorrectly. We were only normalizing the types when
the projection was visited, which could impact on how we identify
whether this structure already existed or not.

This change normalizes all types before starting to codegen the
entire type.