Update Kani copyright blob Amazon -> Kani Contributors
I used the following command to update the blob:

```bash
for f in $(git ls-files); do
  sed -i "0,/Copyright Amazon.com/{s/Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved./Copyright Kani Contributors/}" $f
done
```

To validate the change, I inspected the changes and the list of files
that were modified. I also ran the copyright check script to ensure
that everything was updated.

Finally, I ran the regression.