Fix how we handle niche optimization with ZST
The compiler uses niche optimization when there is only one variant that
contains nonzero sized fields. This variant is represented as the
dataful variant. However, other variants may have one or more ZST fields
that can be deconstructed and referenced in the code.

Before this change, we were not generating code for them and it generate
type mismatch issues when the user tried to access them.

This change ensures that we represent all variants that have any field,
so their access is valid.