Fix how we handle niche optimization with ZST (#1205)
* Fix how we handle niche optimization with ZST

The compiler uses niche optimization when there is only one variant that
contains nonzero sized fields. This variant is represented as the
dataful variant. However, other variants may have one or more ZST fields
that can be deconstructed and referenced in the code.

Before this change, we were not generating code for them and it generate
type mismatch issues when the user tried to access them.

This change ensures that we represent all variants that have any field,
so their access is valid.

* Enable debug assert inside projection mismatch

We should probably turn this into an assert and revert the logic added
by https://github.com/model-checking/kani/pull/1057. But this is a much
larger change that I prefer creating a separate PR for.

Co-authored-by: Zyad Hassan <88045115+zhassan-aws@users.noreply.github.com>