Add support to all object-safe receiver types
For object-safe traits, associated methods that can be dynamically
dispatched must have the receiver type of Self as one of the following:

- &Self
- &mut Self
- Box<Self>
- Rc<Self>
- Arc<Self>
- Pin<P> where P is one of the types above

ref: https://doc.rust-lang.org/reference/items/traits.html#object-safety

Before this change, we only supported &Self and &mut Self. In order to
add support to the other types, which are structures not pointers, we
needed to make sure that the structure type that we encode in the vtable
is compatible with the concrete type (ref 6.2.7.3 ISO/IEC 9899:202x).
For that, we mirror the original type and replace the raw pointer type
to be the data pointer type of `&dyn T`.