Update kani's std library to use rust 2021 (#1352)
* Update kani's std library to use rust 2021

For that, we needed to change how we handle panic messages. In rust
2021, the panic implementation will always invoke panic_fmt(), which we
cannot handle really well. Create our own panic function and use that
instead.

* Ignore std during docs build for now

There seems to be an issue with `rustdoc` that is crashing while trying
to document the std library. Ignore it until this gets fixed.