Improve std-lib-regression script
Add a dummy proof-harness to it and fix the issue with duplicated lang
items that was triggered when trying to resolve the dependencies. The
issues seemed to be related to the fact that `kani` crate was built with
a different version of the `std` library.

For now, we explicitly add the `kani` crate as a dependency so it gets
rebuilt with the fresh std.