Fix `unreachable!()` override macro + new tests
My latest change to handling panic messages actually broke `unreachable`
macro. The main issue was the wrong position of a parenthesis and the
fact that we didn't have tests that covered that specific case of the
macro.

One thing I haven't figured out yet is how to trigger an error if the
user code uses edition 2021+ and has the following construct:

```rust
let msg = "blah";
panic!(msg);
```

This is actually an error in 2021+, but the macro override that we
perform accepts that no matter the edition.

Fixes #1362