Fix static variable malformed symbol (#1380)
A symbol's `base_name` has to be a suffix of its `name` otherwise it is
considered a malformed symbol. For static variables, Kani was using the
`pretty_name` to generate the `base_name` which was generating invalid
iRep as a result. This was caught while trying to run symtab2gb after
the std compilation.

Fixes #1361