Fix compilation of panic!() when the arg is another macro.
In rust, a macro cannot expand another macro that was given as an
argument. However, some of the std macros are expanded inside the
compiler, which gives them a way around that.

Because we implement those macros are regular macros, we cannot expand
other macros and we also cannot limit the first argument to be a
literal.

I decided for now to just relax what we accept as the first argument.
This allow us to accept whatever the std panic!() macro accepts, but it
may accept invalid things as well, such as:

```
  panic!("invalid" / "division");
```

That was already accepted as the second or third arguments, so I don't
think this is big step back. This also won't impact soundness since
whatever issue that could rise from creating the expression is already
inside a panic statement; thus verification has failed already.