Build custom sysroot with `--always-encode-mir` on (#1717)
* Build sysroot and repurpose make-kani-release

This change builds a custom sysroot for Kani. This new sysroot will contain a "lib/" folder with Kani libraries as well as the standard libraries compiled with --always-encode-mir. This enable us to fully traverse the std MIR and fix the missing functions errors.

Other changes to the build were described in the issue here: #1605 (comment)

* Add pre-compiled libraries to the bundle.

The size of the bundle did increase quite a bit on my linux machine. It went from 24MB to 67MB.

This is still pretty far from GitHub's max size of 2GB:
https://docs.github.com/en/repositories/releasing-projects-on-github/about-releases#storage-and-bandwidth-quotas