Fix issue with mutable static from upstream crate
The static allocation function in this case was initializing with a
constant. However, because the static is marked as `mut` the compiler
doesn't inline the value. We need to codegen the static and its
initialization too. The check that the MIR was available returned
`false` for some strange reason.

Since we expect that everything that is not foreign to rust to be
available, we just skip that check.