Fix MIR Linker handling of `--function main`
For the specific case where we run Kani with `--function main`, we
compile the crate as a binary. The logic for public functions in the MIR
linker didn't account for the entry_fn(()) which is set in those cases.

Thus, we explicitly check for the entry function when collecting
starting point and the reachability option is set to public functions.