Disable debug_asserts inside `std` due to poor UX
This is a possible workaround for now to improve the UX around UB
detection in intrinsics when using the MIR Linker. The issue is tracked
here:
https://github.com/model-checking/kani/issues/1740

Basically, the rust toolchain uses a release build that removes all
debug assertions from the standard library. When we switched to using
our custom build of the `std` we decided to enable the debug assertions
in order to catch more potential UBs. However, the UX is not always
great. The assertions don't have any clear descriptions and they may
fail in unexpected places. E.g.: The violation of an intrinsic safety
condition triggers the following failures:
```
> Failed checks: Called `Option::unwrap()` on a `None` value
```