Remove fixme test because it may take a long time
This is an odd one, but I believe this test is causing some workflows to
take hours. It doesn't always happen, and it depends on the machine, but
I think for now we might as well remove this test.

These runs take a long time, and the only test flagged as running above
60s is the one I'm removing now:
- https://github.com/model-checking/kani/actions/runs/3380137117/jobs/5612577106#step:5:238
- https://github.com/model-checking/kani/actions/runs/3348493009/jobs/5547606884#step:5:236