Compiletest improvements (#1848)
This adds `--timeout <val>` option to compiletest so we can add some hard timeout to some of our regression suites. My suggestion is to add timeout values for suites that are required to run in every PR. Suites like `slow_test` and `perf` may or may not have timeout.

I have also added a `--dry-run` that allow us to see the compiletest configuration and print the tests that were found.

For now, the only change I made to the regression is to print the compiletest configuration. This is to help us debug potential CI failures or timing issues.

Co-authored-by: Adrian Palacios <73246657+adpaco-aws@users.noreply.github.com>