Refactor coercion logic for future usage inside codegen (#1837)
We currently have a few issues with how we are generating code for casting (#566, and #1528). The structure of the code is also hard to understand and maintain (see #1531 for more details).

This PR is the first part of the fix I developed. This change moves the coercion specific code to its own module and it introduces an iterator that traverses the coercion path.