Remove github action version check
We currently have a circular dependency.

The github action CI enforces that the tracking version has been
released. So we can't really update the github action until the release
is complete.

So by design, the action repository is always one version behind by the
time we run the release workflow.