Remove github action version check (#1907)
We currently have a circular dependency. The action repository cannot be on the same version as Kani by the time we run the release workflow.

The github action CI enforces that the tracking version has been released. So we can't really update the github action until the release is complete.