Add support to casting closure to function pointer
Change function declaration and call to ignore arguments that are ZST,
and implement the closure to function pointer by reifying it to
its FnOnce implementation.

The casting of closures to function pointers rely on the fact that the
closures do not capture anything. Those closures are represented by
empty structures, which should get pruned from its argument list of the
FnOnce::call_once implementation.