Store failure reason into assess metadata (#2166)
We now store Kani's version and the failure reason (if there's one) into the assess metadata. For now, we still need to manually parse this information, but this moves us closer to #2058 and #2165.

In order to collect more details about the compilation errors, we now inspect the cargo diagnostic messages.