Move attribute handling to `kani_middle` and polish warnings (#2211)
As part of removing one harness restriction from stubbing, we will need to run the rust compiler driver multiple times. However, we want to only produce one metadata file with information about all the compiled artifacts.

Thus, I'm planning to move the harness metadata generation to the KaniCompiler as opposed to the GotoCCtxt. As a first step, I'm moving the attributes handling out of the GotoCCtxt.

The only behavior change in this PR is related to some of the warnings we print. I tried to standardize it a bit (following our new guidelines) and I moved the stub warning to be part of the driver.