Add `--fix-expected` to help batch fix to expected tests
This option will basically override the expected file with Kani's output
for every test that fails the expectation. Developers should still prune
the expected file to ensure only the information that is needed gets
pushed.

My motivation is that I always find myself updating expected tests
manually by rerunning Kani for each failure, piping the output to the
expected file and edit the file after. With this hack, I only need to do
the last step manually.

Since we still fail the test, the files that were updated are listed,
which makes it very easy to track down the changes.

Note that with great power comes great responsibility. Developers still
need to manually edit the expected files before publishing a PR.