Add `benchcomp run` and `benchcomp collate`
`benchmark run` is a command that runs every combination of suite x
variant, and writes their results out to files in a single directory.

`benchcomp collate` reads all the files from that directory and writes
out a single combined file that can be further post-processed.

This commit also adds a regression test, together with regression test
infrastructure. The commit also adds a trivial 'parser' for the test.