Fix build-dev cache (#2280)
After updating our dependencies, running `cargo build-dev` always trigger the recompilation of a few dependencies as well as `kani-driver` and `kani-compiler`.

This seems to be caused by [this change](https://github.com/bytecodealliance/rustix/pull/544) on `rustix` dependency.

The problem here is that we share the same build folder for our tools and our binaries. But they build with slightly different cargo arguments, and this package decided to make their build script very sensitive to changes to their environment

To overcome this, force `cargo build-dev` to use a different folder for the tools build.