Add PR-blocking performance regression run
This commit adds a CI job that runs the `benchcomp` tool on the
performance regression suite, comparing the HEAD of the pull request to
the branch that the PR targets.

The CI job fails if any performance benchmark is more than 10% slower
when run using the HEAD version of Kani, or if any performance benchmark
fails with the HEAD version while passing with the base.