Fix error_on_regression only using last check
This commit ensures that error_on_regression checker checks every one of
the checks specified in the config file. Prior to this commit it would
only check the last one.

The commit also increases the threshold that benchcomp uses to report a
regression. Benchmarks that take less than 10 seconds (previously 2
seconds) are now ignored, and a 50% slowdown (previously 10%) is
considered a regression.