Explicitly pass out_file to benchcomp visualizers
This commit adds an `out_file` to the benchcomp visualizes that emit
output (currently just one visualization, dump_yaml). This is to allow
different visualizations to be written to different output files,
including stdout.

The intention (for a future commit) is that CI will save certain
visualizations as artifacts, which users can then download.