Add markdown_results_table benchcomp visualization
This commit adds a new visualization that writes all results out to a
file as a series of tables, one for each metric.

For each metric, each row comprises the benchmark name, followed by the
values of the metric for each of the variants.

This is an example of the output:

```
  ## runtime

  | Benchmark |  variant_1 | variant_2 |
  | --- | --- |--- |
  | bench_1 | 5 | 10 |
  | bench_2 | 10 | 5 |

  ## success

  | Benchmark |  variant_1 | variant_2 |
  | --- | --- |--- |
  | bench_1 | True | True |
  | bench_2 | True | False |
```