Move common arguments to their own structure (#2449)
This is a pre-work needed for implementing #2440 as a subcommand. Today, the only other subcommand we have is assess and the way it takes arguments is rather broken as explained in #1831.

For the playback sub-command, I expect that we will need at least CargoArgs and CommonArgs to control the build process and to control the process verbosity.