Fix static variable initialization when they have the same value (#2469)
We were incorrectly deduping the allocation of constant static variables. Static variables should each have their own memory allocation, and it should be correctly initialized.

The fix to this issue was basically the change to static_var.rs. The changes to operand.rs was a bit of cleanup to make it explicit that we should only try to dedupe constant allocations.