Improve assess and regression time (#2478)
The per-reachability analysis can have a overhead on the compilation time when assessing a project that has lots of tests. Disable it for now when analyzing tests.

Despite that improvement, assessing the firecracker still takes a long time, even with --only-codegen, due to the number of calls done to goto-cc. Thus, don't use assess here.