Change how we translate Array to GOTO (#2481)
Directly codegen a rust Array into a GOTO array. Before we used to wrap it in a struct.

According to the code documentation, we used to wrap it so we could assign an array to another using the struct copy; however, I didn't see any place where this is needed.

The cprover bindings crate incorrectly returned that an array was not an lvalue type, which maybe is the reason why we thought that was needed before. However, array is an lvalue, just not a modifiable lvalue, according to the C standard.
This simplifies the code and hopefully will help with loop systhesis.