Remove --harness requirement from stubbing (#2495)
The compiler can now generate code for different combinations of harnesses with and without stubs. For that, we implemented a little state machine that may execute the underlying rust compiler multiple times. The first run is always for harnesses that don't have stub, while the subsequent runs are for harnesses with stubs.

Co-authored-by: Zyad Hassan <88045115+zhassan-aws@users.noreply.github.com>