Remove atty and tracing-tree, update hermit-abi (#2581)
This removes dependency on atty, and tracing-tree (which depends on atty). This is in response to this security advisory:

https://rustsec.org/advisories/RUSTSEC-2021-0145

atty is removed by switching to std::io::IsTerminal. tracing-tree is removed by replacing HierarchicalLayer with a regular tracing_subscriber::fmt::layer that directs to stderr.

The PR also updates hermit-abi to 0.3.2 from 0.3.1, in response to 0.3.1 being yanked.

This PR resolves #2580.