Add support to array-based SIMD
Originally, repr(simd) supported only multi-field form. An array
based version was later added and it's likely to become the only
supported way.

The array-based version is currently used in the standard library, and
to implement `portable-simd`. This change adds support to instantiating
and using the array-based version.

We should still add more tests to check if all simd operations that Kani
supports works with the array-based SIMD.