Add support to array-based SIMD (#2633)
Originally, repr(simd) supported only multi-field form. An array based version was later added and it's likely to become the only supported way (https://github.com/rust-lang/compiler-team/issues/621).

The array-based version is currently used in the standard library, and it is used to implement `portable-simd`. This change adds support to instantiating and using the array-based version.